% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/250813_eCOMET_build_V2.R
\name{PLSDAplot}
\alias{PLSDAplot}
\title{PLSDAplot}
\usage{
PLSDAplot(
  mmo,
  color,
  topk = 10,
  outdir,
  normalization = "Z",
  filter_feature = FALSE,
  feature_list = NULL,
  filter_group = FALSE,
  group_list = NULL
)
}
\arguments{
\item{mmo}{The mmo object with feature data and metadata}

\item{color}{A vector of colors for the groups in the plot}

\item{topk}{The number of top features to display in the plot (default: 10)}

\item{outdir}{The output file path for the PLS-DA plot}

\item{normalization}{The normalization method to use for feature data. Options are 'None', 'Log', 'Meancentered', or 'Z' (default: 'Z')}

\item{filter_feature}{Boolean to filter features by feature_list (default: FALSE)}

\item{feature_list}{A vector of feature names to filter (default: NULL)}

\item{filter_group}{Boolean to filter groups by group_list (default: FALSE)}

\item{group_list}{A vector of group names to filter (default: NULL)}
}
\description{
PLSDAplot
}
